put code into seperate files, begining the create of an AI function
