+++ play.py
@@ -1,0 +1,34 @@
+
+selected = (-1,-1)
+
+def RedTurn(click):
+    print("hi")
+    # if selected == (-1,-1): #select square
+    #     selected = clickedSquare
+    # elif selected == click: #unselect square
+    #     selected = (-1,-1)
+    # else:
+    #     res = ck.makeTurn(selected,clickedSquare)
+
+    #     if res["valid"]: #if the move was sucsessful
+    #         #set the new spot equal to the correct color
+    #         newSpot = res["drop"]
+    #         ck.grid[newSpot] = ck.grid[selected] 
+
+    #         #loop through the spots that need deleted
+    #         for spot in res["spotsToRemove"]: 
+    #             grid[spot] = "N"
+            
+    #         #unselect square
+    #         selected = (-1,-1)
+            
+    #         #change the turn
+    #         if res["changeTurn"]:
+    #             turn = otherColor(turn)
+    #         else:
+    #             print("There is a double jump")
+            
+    #     else: #the move was invalid
+            
+    #         #print the error
+    #         print(res["error"])
